+++ sickrage/providers/torrent/iptorrents.py
@@ -56,10 +56,10 @@ class IPTorrentsProvider(TorrentProvider):
     def search(self, search_strings, age=0, ep_obj=None):
         results = []
 
-        freeleech = '&free=on' if self.freeleech else ''
-
         if not self.login():
             return results
+
+        freeleech = '&free=on' if self.freeleech else ''
 
         for mode in search_strings:
             sickrage.srCore.srLogger.debug("Search Mode: %s" % mode)
@@ -98,33 +98,26 @@ class IPTorrentsProvider(TorrentProvider):
         results = []
 
         data = re.sub(r'(?im)<button.+?<[/]button>', '', data, 0)
+
         with bs4_parser(data) as html:
-            if not html:
-                sickrage.srCore.srLogger.debug("No data returned from provider")
-                return results
-
-            if html.find(text='No Torrents Found!'):
-                sickrage.srCore.srLogger.debug("Data returned from provider does not contain any torrents")
-                return results
-
             torrent_table = html.find('table', id='torrents')
-            torrents = torrent_table.find_all('tr') if torrent_table else []
+            torrents = torrent_table('tr') if torrent_table else []
 
             # Continue only if one Release is found
-            if len(torrents) < 2:
+            if len(torrents) < 2 or html.find(text='No Torrents Found!'):
                 sickrage.srCore.srLogger.debug("Data returned from provider does not contain any torrents")
                 return results
 
-            for result in torrents[1:]:
+            for torrent in torrents[1:]:
                 try:
-                    title = result.find_all('td')[1].find('a').text
-                    download_url = self.urls['base_url'] + result.find_all('td')[3].find('a')['href']
-                    size = convert_size(result.find_all('td')[5].text, -1)
-                    seeders = int(result.find('td', attrs={'class': 'ac t_seeders'}).text)
-                    leechers = int(result.find('td', attrs={'class': 'ac t_leechers'}).text)
-
+                    title = torrent('td')[1].find('a').text
+                    download_url = self.urls['base_url'] + torrent('td')[3].find('a')['href']
                     if not all([title, download_url]):
                         continue
+
+                    size = convert_size(torrent('td')[5].text, -1)
+                    seeders = int(torrent.find('td', attrs={'class': 'ac t_seeders'}).text)
+                    leechers = int(torrent.find('td', attrs={'class': 'ac t_leechers'}).text)
 
                     item = {'title': title, 'link': download_url, 'size': size, 'seeders': seeders,
                             'leechers': leechers, 'hash': ''}